flask_server_funko_returns_portal/requirements.txt
arrow==0.17.0
binaryornot==0.4.4
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Mail==0.9.1
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.5.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-slugify==4.0.1
requests==2.24.0
six==1.15.0
text-unidecode==1.3
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
